Fix transaction pathing for methods with no ACL guard
I forgot to cover the case where a function is not guarded by
the ACL, which would error out the transaction pathing.

For now, we just assume we can execute actions directly if they
do not have any ACL requirements, but for good measure we
should probably always perform a call to see if the transaction
would succeed.