wrapper: Use memory cache adapter outside browser
So, the primary issue with working in both the browser
and outside the browser right now is to make sure that
we do not require the filesystem specific adapters for lowdb
in the browser, as that will break.

Rather than spend a lot of time on fixing this right now
(which I have already done), we're just going to use the
memory adapter.

This works fine for things like Aragon CLI, but will be a pain
when we start using Electron, unless Electron also supports
localStorage (which I think it does).