feat(wrapper): introduce getRecommendedGasLimit util
In order to implement #150 this commit introduces a new `getRecommendedGasLimit()` util
which is calculated based on the estimatedGas by web3 and the latest block
available gas limit.

This is similar to how MetaMask handles gas limit proposals.

We now also prefer using the provided gas that is passed to an intent, if it's higher
than the recommended gas limit. In other words, provided gas is treated as
minimum gas required to perform the transaction/intent.

Closes #150