Simplify the initAcl logic
Retrieve all events and there by drastically simplify the method.
Because the Acl contract only emits the two events currently,
this doesn't introduce any network overhead for extra data.

Because the events of the two types we're interested in are now
ordered in the stream, we no longer need the added skipWhile
operator and the eventSet passed along