wrapper: deduplicate async requests when fetching app information (#257)
Adds a `AsyncRequestCache` that deduplicates async requests and caches the first successful result. This is really useful for getting information that will stay constant, such as contract values that can never change or contents from ipfs hashes.

Refactors the wrapper's `initApps()` to use this cache and also set up the stage for getting aragonPM repository information.