Forwarded Actions: reshape as object + id arrays
Rather than having sub-keys for `pending`, `completed`, and `failed`
with all information within these structures, this stores all data
within an `actions` key, and keeps separate arrays of
`pendingActionKeys`, `failedActionKeys`, and `completedActionKeys`.

Since actionIds could repeat between different originating apps (for
example, Voting and Dot Voting could each forward separate actions to
Allocations, both with id=1), these keys are a combination of the
currentApp and actionId.

A React component which uses this new data structure will have a
parallel and easy-to-use interface for whichever sort of forwarded
actions it wishes to show, such as:

    import { useAragonApi } from '@aragon/api-react'
    import { Pending, Failed, Generic } from '.'

    function MyComponent () {
      const { forwardedActions: forwarded } = useAragonApi()

      return [
        <h1>Pending</h1>,
        forwarded.pendingActionIds.map(id => (
          <Pending key={id} {...forwarded.actions[id]} />
        )),

        <h1>Recently Failed</h1>,
        forwarded.failedActionIds.slice(-10).map(id => (
          <Failed key={id} {...forwarded.actions[id]} />
        )),

        <h1>All Pending and Failed</h1>,
        [
          ...forwarded.failedActionIds,
          ...forwarded.pendingActionIds
        ].map(id => (
          <Generic {...forwarded.actions[id]} />
        ))
      ]
    }

If a component wishes to break down pending actions per-forwarder, this
new interface makes it possible to do so, as long as aragonAPI also
provides a list of all installed apps. This might look something like:

    import { useAragonApi } from '@aragon/api-react'
    import { Pending } from '.'

    function MyComponent () {
      const { forwardedActions: forwarded, installedApps } = useAragonApi()

      const pendingActions = forwarded.pendingActionIds.map(
        id => forwarded.actions[id]
      )

      return Array.from(
        new Set(pendingActions.map(a => a.currentApp))
      ).map(appId => [
        <h1>Pending for {installedApps[appId]}</h1>,
        pendingActions.filter(a => a.currentApp === appId).map(action =>
          <Pending key={action.actionId} {...action} />
        )
      ])
    }