setForwardedAction: ensure well-formed arguments
This code uses the approach of throwing traditional errors within
`setForwardedAction`, rather than throwing in an RxJS-compatible way
from within the `scan` function. This required minor modifications to
the argument function for `forwardedActions.next`. Given that this
function is meant to be private, this should not break any consuming
code.