forwardedActions/cache: fix false positive test
Due to the confusion of code mutation, the original test caused the
code-under-test to update the in-memory object with name `babaStored`.
This meant that `babaStored` also had a key for `cacaAddress`,
`dadaAddress`, and `fafaAddress`.

This new code is significantly smarter about dealing with in-memory
object mutation, both in the test and in the code itself.

As the new tests demonstrate, `cachedActions` is stored in the JS
runtime's memory for as long as the process runs, which means it can
continue to be updated throughout the lifetime of that process. Care
must be taken to ensure that `cachedActions` and `inMemoryActions` do
not reference the same location in memory, neither at the root nor at
any leaf.

This whole thing feels strange, and is certainly harder to read than
immutable code. If this does, in fact, prevent unnecessary re-renders in
any consumers that use forwardedActions, then I think the oddity is
worth the hassle. With the extensive tests, I feel reasonably confident
that this code does behave as intended.